hamiltonian implemented
